Update PutServ & PutHelp to PutNow
This sends text to the server immediately, bypassing all queues. Use with caution, as the bot may easily flood itself off the server.

Putnow is only available on EggDrop as far as I am aware, so if you're using other forms of TCL emulator, you may not want to merge this.